// src/components/modals/guitar/types.ts
export interface GuitarLesson {
  title: string;
  artist: string;
  key: string;
  tuning: string;
  lyricsWithChords: string;
  chordChanges: string;
  inspiration: string;
  youtubeLessonTitle: string;
  youtubeLessonUrl: string;
  spotifyUrl: string;
}

export interface GuitarSong {
  title: string;
  artist: string;
}